Bye bye ccache
Just one run through the guide generates ~2 GiB of ccache and overall
just slows down the build.

Since it's more likely that the build will be done rarely rather than
often, it really doesn't make sense to promote this. It can still of
course be enabled manually if really needed 🤷